requirements.txt
wheel==0.35.0
absl-py==0.13.0
astunparse==1.6.3
autopep8==1.5.7
cachetools==4.2.2
catkin-pkg==0.4.23
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
dlib==19.22.0
docopt==0.6.2
docutils==0.17.1
dynamixel-sdk==3.7.31
empy==3.3.4
ffmpeg-python==0.2.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
gTTS==2.2.3
h5py==3.1.0
idna==3.2
ikpy==3.2.2
imutils==0.5.4
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lark-parser==0.11.3
Markdown==3.3.4
mpmath==1.2.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycodestyle==2.7.0
pydub==0.25.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
pyserial==3.5
python-dateutil==2.8.2
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rptree==0.1.1
rsa==4.7.2
scipy==1.7.1
six==1.15.0
SpeechRecognition==3.8.1
sympy==1.8
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tf==1.0.0
tflearn==0.5.0
toml==0.10.2
tqdm==4.62.0
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
yarg==0.1.9
